Add Battle.Net Services
This can probably be cleaned up a bit, but I wasn't sure how.  Each
region uses the same syntax and pattern, however all URIs have a
different sub-domain.  I implemented it as a base class and overrides
for the region, but ideally it should be a single class that can take a
region setting.  I wasn't sure how to tackle such a thing consistent
with the architectural objectives of the project though.